+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Changelog
 
 ### 2.0.0
  - Updated to use Thymeleaf 3
+ - Change `RssXmlGenerator.generate()` method to accept just the channel and
+   output - items to publish are now part of the channel.
 
 ### 1.0.0
  - Initial release